# docs/swagger.yaml
basePath: /v1
definitions:
  handler.httpError:
    properties:
      message:
        example: Some error message
        type: string
    type: object
...
  request.RegisterReq:
    properties:
      email:
        example: example@economicus.kr
        type: string
      nickname:
        example: user nickname
        type: string
      password:
        example: some password
        type: string
    type: object
...
  response.ProfileQuantResponse:
    properties:
      chart:
        example:
        - 8.31201046811529
        - 15.13554790878776
        - -1.336521221573761
        - -1.42408166715555
        - 10.420784591586559
        - 8.305691643668455
        - 17.68356243256443
        - 9.407034979656027
        - -4.15162926200139
        - 5.542443496088845
        - 6.654446258518339
        items:
          type: number
        type: array
      quant_id:
        example: 5
        type: integer
    type: object
  response.ProfileResponse:
    properties:
      quant:
        items:
          $ref: '#/definitions/response.ProfileQuantResponse'
        type: array
      user:
        $ref: '#/definitions/table.User'
    type: object
  table.Location:
    properties:
      city:
        type: string
      country:
        type: string
    type: object
  table.Profile:
    properties:
      email:
        type: string
      intro_message:
        type: string
      location:
        $ref: '#/definitions/table.Location'
      nickname:
        type: string
      phone:
        type: string
      profile_image:
        type: string
      user_id:
        type: integer
      user_url:
        type: string
    type: object
  table.User:
    properties:
      email:
        type: string
      profile:
        $ref: '#/definitions/table.Profile'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Economicus 메인 서버
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.economicus.kr/
  title: Economicus Main Server
  version: 1.0.0
paths:
...
  /register:
    post:
      consumes:
      - application/json
      description: 이메일, 비밀번호로 유저 회원가입
      parameters:
      - description: A user information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.httpError'
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Register a user
      tags:
      - user
  /users:
    get:
      consumes:
      - application/json
      description: 이메일, 비밀번호로 유저 회원가입
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: A user information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.httpError'
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Return all users
      tags:
      - user
  /users/profile/{user_id}:
    get:
      consumes:
      - application/json
      description: 프로필 화면에서 유저의 정보 및 보유한 퀀트 모델을 반환
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: User id to get profile
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Profile response
          schema:
            $ref: '#/definitions/response.ProfileResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.httpError'
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Return user's profile and quants
      tags:
      - user
schemes:
- https
- http
securityDefinitions:
  JWT:
    type: basic
swagger: "2.0"